wire up unit tests
